requirements.txt
flask==3.0.2
pillow==11.2.1
img2pdf==0.5.1
gunicorn==21.2.0
werkzeug==3.0.1
python-dotenv==1.0.1
python-uno==2.5.1 